Undecorate game-factory, rename to game-engine
