+++ lib/debug-event-logger.js
@@ -4,14 +4,8 @@ const bus = require('./message-bus');
 
 const style = require('./style');
 
-class DebuggingGameFactory {
-  constructor(gameFactory) {
-    this._gameFactory = gameFactory;
-  }
-
-  createGame() {
-    this._gameFactory.createGame();
-
+class DebugEventLogger {
+  static subscribe() {
     bus.eventChannel.subscribe('#', (data, envelope) => {
       switch (envelope.topic) {
         case 'game.started':
@@ -74,4 +68,4 @@ function formatNode(node) {
   return `${node.name}(${node.id})[${node.location.x},${node.location.y},${node.location.z}]`;
 }
 
-module.exports = DebuggingGameFactory;
+module.exports = DebugEventLogger;
