+++ lib/index.js
@@ -8,7 +8,7 @@ const GameDefinitionRepository = require('./game-definition-repository');
 const MapNodeRepository = require('./map-node-repository');
 const ItemRepository = require('./item-repository');
 
-const DebuggingGameFactory = require('./debugging-game-factory');
+const DebugEventLogger = require('./debug-event-logger');
 const GameFactory = require('./game-factory');
 
 const opt = getopt.create([
@@ -66,11 +66,11 @@ new StartGameHandler(repositorySet).subscribe();
 new StopGameHandler().subscribe();
 new ListInventoryHandler().subscribe();
 
-let gameFactory = new GameFactory(repositorySet);
-
 if (opt.options.debug) {
-  gameFactory = new DebuggingGameFactory(gameFactory);
+  DebugEventLogger.subscribe();
 }
+
+const gameFactory = new GameFactory(repositorySet);
 
 const textEngine = new TextEngine(gameFactory, commandParsers);
